ci: fix prettier configuration
In a past change, we configured ESLint to run Prettier using
eslint-plugin-prettier, but we didn't add a Prettier configuration file.
We didn't notice that the Prettier configuration file was missing
because that change didn't trigger the linting process over the whole
codebase.

Additionally, set VALIDATE_TYPESCRIPT_PRETTIER=true. Although this results
in Prettier running two times (using ESLint as described above, and
standalone), this will help when we'll enable "fix mode", so we can set
the corresponding FIX_TYPESCRIPT_PRETTIER variable to "true".

Change-Id: I23d46a18eaae8beb73a13cbc18c6e82aefd49a87
GitOrigin-RevId: 98b07e231b8c15d2f552293d36079b84b3de9b9d